third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-remoteDescription-promise-expected.txt

Tests RTCPeerConnection remoteDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.setRemoteDescription().catch(expectedTypeError) did not throw exception.
PASS pc.setRemoteDescription(null).catch(expectedInvalidSessionDescription) did not throw exception.
PASS pc.setRemoteDescription(sessionDescription).then(requestSucceeded1, requestFailed1); did not throw exception.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS error.name is "OperationError"
PASS expectedInvalidSessionDescription was called.
PASS requestSucceeded was called.
PASS pc.setRemoteDescription(sessionDescription).then(requestSucceeded2, requestFailed2); did not throw exception.
PASS requestFailed was called.
PASS pc.remoteDescription.type is "answer"
PASS pc.remoteDescription.sdp is "remote"
PASS pc.remoteDescription.type is "answer"
PASS pc.remoteDescription.sdp is "remote"
PASS successfullyParsed is true

TEST COMPLETE
